refactor(auth): move it to middlewares package
