+++ adk/server/middlewares/auth.go
@@ -1,4 +1,4 @@
-package server
+package middlewares
 
 import (
 	"context"
